+++ src/api/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - Upgrade setuptools to `75.5.0`
 - Send DB query details on Statique API errors only in debug mode
 - Move `num_pdl` field to a 64-chars string
+- Return created objects UUIDs for statuses and sessions
 
 ## [0.14.0] - 2024-11-15